+++ demosales-domain/sales/src/main/java/com/imecatro/demosales/domain/sales/add/usecases/GetMostPopularProductsUseCase.kt
@@ -5,7 +5,7 @@ import com.imecatro.demosales.domain.sales.add.repository.AddSaleRepository
 class GetMostPopularProductsUseCase(
     private val addSaleRepository: AddSaleRepository
 ) {
-    suspend operator fun invoke() : List<Int>{
+    suspend operator fun invoke() : List<Long>{
        return addSaleRepository.filterPopularProducts(10)
     }
 }
